requirements.txt
fastapi
pandas
python-dotenv
uvicorn
